move section to common component
